srv/requirements.txt
pyyaml>=3.11
requests==0.11.1
grequests==0.2.0

